News Analytics App/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.127
botocore==1.12.127
bs4==0.0.1
bz2file==0.98
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
docutils==0.14
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
funcy==1.12
future==0.17.1
gensim==3.7.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
lxml==4.3.3
MarkupSafe==1.1.1
more-itertools==7.0.0
nltk==3.4
numexpr==2.6.9
numpy==1.16.0
pandas==0.24.0
Pillow==6.0.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
pyLDAvis==2.1.2
PyQt5==5.11.3
PyQt5-sip==4.19.13
pytest==4.4.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
soupsieve==1.9
SQLAlchemy==1.3.2
urllib3==1.24.1
Werkzeug==0.15.2
WTForms==2.2.1
